civic-extract/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pytesseract==0.3.10
pdfplumber==0.10.3
pandas==2.1.3
opencv-python==4.8.1.78
pillow==10.1.0
beautifulsoup4==4.12.2
python-jose==3.3.0